requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
Flask-Migrate
Flask-RESTful
python-dotenv
